% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_05_pips_rcpp_brewer}
\alias{boot_05_pips_rcpp_brewer}
\title{Barbiero and Mecatti 0.5 \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS bootstrap}
\usage{
boot_05_pips_rcpp_brewer(x_sample, x_pop_total, n_boots)
}
\arguments{
\item{x_sample}{Numeric vector. Positive values of an auxiliary variable for
sample elements.}

\item{x_pop_total}{Number. Sum of the auxiliary variable in the
population.}

\item{n_boots}{Integer. A number of bootstrap samples to be generated.}
}
\value{
Integer matrix with \code{n_boots} rows and \code{length(x_sample)}
columns. Each row is a bootstrap sample, and the number in i-th column
denotes the number of times i-th unit from the original sample occurs in
the bootstrap sample.
}
\description{
The function generates bootstrap samples using Barbiero and Mecatti
(2010, p. 62) "0.5 \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS algorithm".
Uses Brewers' design for sampling.
}
\examples{
boot_05_pips_rcpp(1:3, 100, 10, UPbrewer_rcpp)
boot_05_pips_rcpp(1:10, 1000, 5, UPbrewer_rcpp)
}
\references{
Barbiero A. and Mecatti F. (2010). Bootstrap algorithms for
variance estimation in \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS sampling,
in: Complex Data Modeling and Computationally Intensive Statistical
Methods, ed. P. Mantovan and P. Secchi, Springer, Milano.
}
