% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_AT2011_alg3_rcpp}
\alias{boot_AT2011_alg3_rcpp}
\title{Antal and Tille (2011) bootstrap for simple random sampling without replacement}
\usage{
boot_AT2011_alg3_rcpp(N, n, n_boots)
}
\arguments{
\item{N}{Integer. Original population size.}

\item{n}{Integer. Sample size.}

\item{n_boots}{Integer. Number of bootstrap samples to be generated.}
}
\value{
Integer matrix with \code{n_boots} rows and \code{n} columns.
Each row is a bootstrap sample, with entries in columns denoting how many
times a given unit is resampled.
}
\description{
This function implements method of resampling described in the algorithm 3 of
Antal & Tille (2011), p. 539.
}
\examples{
boot_AT2011_alg3_rcpp(10, 5, 20)

}
\references{
Erika Antal and Yves Tille (2011) A Direct Bootstrap Method for
Complex Sampling Designs From a Finite Population, Journal of the American
Statistical Association, 106:494, 534-543, DOI: 10.1198/jasa.2011.tm09767
}
