% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_AT2011_alg45_rcpp}
\alias{boot_AT2011_alg45_rcpp}
\title{Antal and Tille (2011) bootstrap for complex sampling designs}
\usage{
boot_AT2011_alg45_rcpp(pi_sample, n_boots, scheme)
}
\arguments{
\item{pi_sample}{Numeric vector. Inclusion probabilities for elements in the original sample.}

\item{n_boots}{Integer. A number of bootstrap samples to be generated.}

\item{scheme}{Function. Implementation of a sampling scheme (takes a vector
of inclusion probabilities as an argument and returns a sample: a vector of
0's and 1's of the same length as the argument) used to select the original sample.}
}
\value{
Integer Matrix with \code{n_boots} rows and \code{length(pi_sample)}
columns. Each row is a bootstrap sample with entries denoting how many
times given unit is resampled.
}
\description{
Function implements a method of resampling from a sample selected with unequal
probabilities without replacement as described in algorithms 4 and 5 in
Antal & Tille (2011), pp. 539-540.
}
\examples{
pi_sample <- inclusion_probabilities_rcpp(1:10, 6)
boot_AT2011_alg45_rcpp(pi_sample, 5, UPbrewer_rcpp)

}
\references{
Erika Antal and Yves Tille (2011) A Direct Bootstrap Method for
Complex Sampling Designs From a Finite Population, Journal of the American
Statistical Association, 106:494, 534-543, DOI: 10.1198/jasa.2011.tm09767
}
