% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_AT2014_rcpp}
\alias{boot_AT2014_rcpp}
\title{Antal and Tille (2014) bootstrap for complex sampling designs}
\usage{
boot_AT2014_rcpp(pi_sample, n_boots)
}
\arguments{
\item{pi_sample}{Numeric vector of inclusion probabilities of the first kind
(of elements in the original sample).}

\item{n_boots}{Integer. Number of bootstrap samples to create.}
}
\value{
Integer matrix with \code{n_boots} rows and \code{length(phik)} columns.
Each row is a bootstrap sample, with entries in columns denoting how many
times given unit is resampled.
}
\description{
This function implements resampling algorithm described in Antal and Tille
(2014), pp. 1355-1356 for sampling with unequal probabilities without replacement.
}
\references{
Erika Antal and Yves Tille (2014) A new resampling method for
sampling designs without replacement: the doubled half bootstrap.
Comput Stat 29, 1345-1363, https://doi.org/10.1007/s00180-014-0495-0.
}
