% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_Hm_pips.R
\name{boot_Hm_pips}
\alias{boot_Hm_pips}
\title{Modified Holmberg \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS bootstrap}
\usage{
boot_Hm_pips(x_sample, x_pop_total, n_boots, scheme)
}
\arguments{
\item{x_sample}{Numeric vector. Positive values of an auxiliary variable for sample elements.}

\item{x_pop_total}{Number. Sum of the auxiliary variable in the population.}

\item{n_boots}{Integer. A number of bootstrap samples to be generated.}

\item{scheme}{Function. Implementation of a sampling scheme (takes a vector
of inclusion probabilities as an argument and returns a sample: a vector of
0's and 1's of the same length as argument).}
}
\value{
Integer matrix with \code{n_boots} rows and \code{length(x_sample)}
columns. Each row is a bootstrap sample, and the number in i-th column
denotes the number of times i-th unit from the original sample occurs in
the bootstrap sample.
}
\description{
The function generates bootstrap samples using Holmberg (1998, p. 380)
bootstrap \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS sampling modified such that
the generation of the pseudo-population occurs in each bootstrap iteration.
}
\examples{
boot_Hm_pips(1:3, 100, 10, UPbrewer_rcpp)
boot_Hm_pips(1:10, 1000, 4, UPbrewer_rcpp)

}
\references{
Holmberg, A.(1998). A bootstrap approach to probability proportional
to size sampling. \emph{Proceedings of Section on Survey Research Methods}.
American Statistical Association, Washington, 378-383.
}
