% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_Q_pips_rcpp}
\alias{boot_Q_pips_rcpp}
\title{Quatember \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS bootstrap}
\usage{
boot_Q_pips_rcpp(x_sample, x_pop_total, n_boots)
}
\arguments{
\item{x_sample}{Numeric vector. Positive values of an auxiliary variable for sample elements.}

\item{x_pop_total}{Numeric. Sum of the auxiliary variable in the population.}

\item{n_boots}{Integer. A number of bootstrap samples to be generated.}
}
\value{
Integer matrix with \code{n_boots} rows and \code{length(x_sample)}
columns. Each row is a bootstrap sample, and the number in i-th column
denotes how many times a given unit is resampled.
}
\description{
The function generates bootstrap samples using Quatember (2014, p. 95)
bootstrap \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS sampling.
}
\examples{
boot_Q_pips_rcpp(1:3, 100, 10)
boot_Q_pips_rcpp(1:10, 1000, 4)

}
\references{
Quatember, A. (2014). The Finite Population Bootstrap - From the
Maximum Likelihood to the Horvitz-Thompson Approach. Austrian Journal of
Statistics, 43(2), 93-102. https://doi.org/10.17713/ajs.v43i2.10
}
