% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_Qg_pips_rcpp}
\alias{boot_Qg_pips_rcpp}
\title{Generalized Quatember \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS bootstrap}
\usage{
boot_Qg_pips_rcpp(x_sample, x_pop_total, g_weights, n_boots)
}
\arguments{
\item{x_sample}{Numeric vector. Positive values of an auxiliary variable for sample elements.}

\item{x_pop_total}{Number. Sum of the auxiliary variable in the population.}

\item{g_weights}{Numeric vector of the length \code{length(x_sample)}. Positive
weights for sample elements. Should sum up to the sample size to ensure that
products of 'x_sample' and calibration weights (defined as products of
'g_weights` and inverses of first order inclusion probabilities) are equal
to 'x_pop_total'. If this condition is not met, they are rescaled.}

\item{n_boots}{Integer. A number of bootstrap samples to be generated.}
}
\value{
Integer matrix with \code{n_boots} rows and \code{length(x_sample)}
columns. Each row is a bootstrap sample, with entries in columns denoting
how many times given unit is resampled.
}
\description{
The function generates bootstrap samples using generalized Quatember (2014, p. 95)
bootstrap \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS sampling where additional weights
are taken into account to compute the number of replicates in the pseudo-population.
}
\examples{
N <- 200 # population size
x_population <- rgamma(N, 1)
n <- 10 # sample size
pi <- inclusion_probabilities_rcpp(x_population, n)
set.seed(123)
sample <- UPbrewer_rcpp(pi)
x_s <- x_population[sample == 1]
pi_s <- pi[sample == 1]
# g-weigths
g <- sampling::calib(Xs = cbind(x_s, rep(1, n)),
                     d = 1 / pi_s,
                     total = c(sum(x_population), N),
                     method = "truncated",
                     bounds = c(low = 0, upp = 10))
boot_Qg_pips_rcpp(x_s, sum(x_population), g, 5)

}
\references{
Quatember, A. (2014). The Finite Population Bootstrap - From the
Maximum Likelihood to the Horvitz-Thompson Approach. Austrian Journal of
Statistics, 43(2), 93-102. https://doi.org/10.17713/ajs.v43i2.10

Zadlo T. (2021) On the generalisation of Quatember's
bootstrap, Statistics in Transition new series, 22(1), 163-178,
https://doi.org/10.21307/stattrans-2021-009.
}
