% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_SP_pips_rcpp}
\alias{boot_SP_pips_rcpp}
\title{Sverchkov and Pfeffermann \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS bootstrap}
\usage{
boot_SP_pips_rcpp(x_sample, pi_sample, N, n_boots, scheme)
}
\arguments{
\item{x_sample}{Numeric vector. Positive values of an auxiliary variable for sample elements.}

\item{pi_sample}{Numeric vector. Values of inclusion probabilities for sample elements.}

\item{N}{Integer. Population size.}

\item{n_boots}{Integer. A number of bootstrap samples to be generated.}

\item{scheme}{Function. Implementation of a sampling scheme (takes a vector of inclusion probabilities as an argument and returns a sample: a vector of 0's and 1's of the same length as argument).}
}
\value{
Integer matrix with \code{n_boots} rows and \code{length(x_sample)} columns. Each row is a bootstrap sample, and the number in i-th column denotes the number of times i-th unit from the original sample occurs in the bootstrap sample.
}
\description{
The function generates bootstrap samples using Sverchkov and Pfeffermann
(2004, p. 85) bootstrap \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS sampling
}
\examples{
boot_SP_pips_rcpp(1:3, c(0.005, 0.01, 0.015), 100, 10, UPbrewer_rcpp)
boot_SP_pips_rcpp(1:10, (1:10) * 2e-5, 1000, 4, UPbrewer_rcpp)

}
\references{
Sverchkov M. and Pfeffermann D. (2004). Prediction of Finite Population Totals Based on the Sample Distribution, \emph{Survey Methodology}, 30(1):79-92.
}
