% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_xbal1_pips_rcpp_brewer}
\alias{boot_xbal1_pips_rcpp_brewer}
\title{Barbiero and Mecatti x-balanced 1 \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS bootstrap}
\usage{
boot_xbal1_pips_rcpp_brewer(x_sample, x_pop_total, n_boots)
}
\arguments{
\item{x_sample}{Numeric vector. Positive values of an auxiliary variable for sample elements.}

\item{x_pop_total}{Number. Sum of the auxiliary variable in the population.}

\item{n_boots}{Integer. A number of bootstrap samples to be generated.}
}
\value{
Integer matrix with \code{n_boots} rows and \code{length(x_sample)}
columns. Each row is a bootstrap sample, and the number in i-th column
denotes how many times a given unit is resampled.
}
\description{
The function generates bootstrap samples using Barbiero and Mecatti
(2010, p. 62-63) x-balanced 1 bootstrap \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS sampling
}
\examples{
boot_xbal1_pips_rcpp_brewer(1:3, 100, 10)
boot_xbal1_pips_rcpp_brewer(1:10, 1000, 4)

}
\references{
Barbiero A. and Mecatti F. (2010). Bootstrap algorithms for
variance estimation in \ifelse{html}{\out{&pi;}}{\eqn{\pi}}PS sampling,
in: Complex Data Modeling and Computationally Intensive Statistical
Methods, ed. P. Mantovan and P. Secchi, Springer, Milano.
}
