% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate statistics based on bootstrap estimates}
\usage{
calculate_stats(estimator_value, bootstrap_estimates, confidence_level)
}
\arguments{
\item{estimator_value}{Numeric. Value of estimator based on a sample}

\item{bootstrap_estimates}{Vector. Values of estimator based on bootstrap resamples}

\item{confidence_level}{Numeric. Confidence level used for calculations of confidence interval limits.}
}
\value{
List with the following elements:
\itemize{
\item LCL_boot, UCL_boot: lower and upper confidence limits based on quantiles in bootstrap distribution,
\item LCL_normboot, UCL_normboot: lower and upper confidence limits based on normal distribution and bootstrap SE estimator
\item estVar, estSE: bootstrap variance and standard error of the estimator.
\item est: estimator value (the first argument).
}
}
\description{
Calculate statistics based on bootstrap estimates
}
\examples{
calculate_stats(rnorm(1), rnorm(1000), 0.95)

}
