% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{halfdoubled_rcpp}
\alias{halfdoubled_rcpp}
\title{Doubled half sampling}
\usage{
halfdoubled_rcpp(n)
}
\arguments{
\item{n}{Integer. Number of elements to sample from.}
}
\value{
Integer vector of size \code{n}. Values indicates how many times
units are selected into sample, 0 means that the unit is not
in the sample.
}
\description{
Draws a sample using "doubled half sampling" described in
Antal and Tille (2014), pp. 1351-1352. Samples are drawn with replacement.
}
\examples{
halfdoubled_rcpp(5)
halfdoubled_rcpp(20)

}
\references{
Antal, E. and Tille, Y. (2014) A new resampling method for
sampling designs without replacement: the doubled half bootstrap.
Comput Stat 29, 1345-1363, https://doi.org/10.1007/s00180-014-0495-0
}
