% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inclusion_probabilities_rcpp}
\alias{inclusion_probabilities_rcpp}
\title{Inclusion probabilities}
\usage{
inclusion_probabilities_rcpp(values, size)
}
\arguments{
\item{values}{Numeric vector. Positive values of an auxiliary variable.}

\item{size}{Integer. Sample size.}
}
\value{
Vector of size \code{length(values)} of inclusion probabilities.
}
\description{
Calculates first order inclusion probabilities for a probability
proportional-to-size sampling design based on positive values of an
auxiliary variable. We use the algorithm described
in Tille (2006), pp. 18-19.
}
\examples{
inclusion_probabilities_rcpp(1:10, 5)
inclusion_probabilities_rcpp(1:10, 6)
inclusion_probabilities_rcpp(c(1:8, 50, 100), 5)

}
\references{
Tille, Y. (2006) Sampling algorithms, Springer, New York.
}
