% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inverse_hypergeometric_rcpp_scalar}
\alias{inverse_hypergeometric_rcpp_scalar}
\title{Inverse Hypergeometric random numbers}
\usage{
inverse_hypergeometric_rcpp_scalar(N, n)
}
\arguments{
\item{N}{Positive integer}

\item{n}{Vector of nonnegative integers.}
}
\value{
Integer.
}
\description{
Generates random number from a special case of the inverse hypergeometric distribution.
}
\details{
Generates random number from a discrete distribution with the probability mass function:
\deqn{\mathbb{P}(X = j) = \dfrac{\dbinom{N + n - j - 2}{n - j}}{\dbinom{N + n - 1}{n}} \qquad j = 0,\dots,n,}
being a special case of the inverse hypergeometric distribution.

This distribution may be described as follows: we have \code{n} red balls and \code{N-1} black balls.
We draw balls without replacement, and \code{X} is the number of red balls drawn before first black ball.

The expectation of this distribution is \eqn{\mathbb{E}(X) = n/N}
and variance \eqn{\mathrm{Var}(X) = \dfrac{n(N-1)(N+n)}{N^2(N+1)}}.
Corner cases are covered in the following way:
if \eqn{n = 0} then \eqn{X \equiv 0} and if \eqn{N = 1} then \eqn{X \equiv n}.
}
\examples{
inverse_hypergeometric_rcpp_scalar(10, 20)

}
