% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampledata_bootrep_rcpp}
\alias{sampledata_bootrep_rcpp}
\title{Bootstrap replications of sample data}
\usage{
sampledata_bootrep_rcpp(bootstrap_samples, x_sample)
}
\arguments{
\item{bootstrap_samples}{Integer matrix with \code{B} rows and \code{n} columns,
where \code{B} is the number of bootstrap replications and \code{n} is the sample
size. Each row is a bootstrap sample, with entries in columns denoting how
many times given unit is resampled.}

\item{x_sample}{Numeric vector. Values of the variable for sampled elements.}
}
\value{
Matrix with \code{B} rows and \code{n} columns. Each row is a bootstrap sample of values of \code{x_sample}.
}
\description{
The function computes a matrix with bootstrap replications of sample data.
}
\examples{
x_sample <- 1:10
set.seed(111)
bootstrap_samples <- boot_H_pips(x_sample, 1000, 4, UPbrewer_rcpp)
sampledata_bootrep_rcpp(bootstrap_samples, x_sample)

}
