% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampling_oneone_rcpp}
\alias{sampling_oneone_rcpp}
\title{One-one resampling design}
\usage{
sampling_oneone_rcpp(n)
}
\arguments{
\item{n}{Integer. Sample size.}
}
\value{
Integer vector of size \verb{n. The value of }i\verb{-th element tells how many times }i`-th unit appears in the sample.
}
\description{
Implements sampling design described in
Antal and Tille (2011, page 537, Algorithm 2).
}
\details{
Fixed sample size resampling design with replacement. Citing
Antal and Tille (2011) p. 537: "an ad hoc construction used to randomly
select \code{n} units from a sample of size \code{n} in such a way that the
expectation and the variance of \eqn{S^*_k} are equal to 1 (...). This
sampling design is a mixture between a simple random sampling with
replacement and a simple random sampling with  over-replacement".
}
\examples{
sampling_oneone_rcpp(10)

}
\references{
Erika Antal and Yves Tille (2011),
A Direct Bootstrap Method for Complex Sampling Designs From a Finite
Population, Journal of the American Statistical Association 106:494,
534-543, DOI: 10.1198/jasa.2011.tm09767
}
