% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srswor_rcpp_01}
\alias{srswor_rcpp_01}
\title{Simple Random Sampling Without Replacement}
\usage{
srswor_rcpp_01(N, n)
}
\arguments{
\item{N}{Integer. Population size - number of elements to draw from.}

\item{n}{Integer. Sample size - number of elements to draw.}
}
\value{
Integer vector of size \code{N} with exactly \code{n} elements
equal to 1 (units selected into the sample) and others equal to 0 (units not
selected).
}
\description{
Function draws a simple random sample without replacement. This is an Rcpp
replacement for \code{\link[sampling]{srswor}}, however with different
implementation.
}
\examples{
srswor_rcpp_01(10, 5)

}
