% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srswr_rcpp_zero}
\alias{srswr_rcpp_zero}
\title{Simple Random Sampling With Replacement}
\usage{
srswr_rcpp_zero(N, n)
}
\arguments{
\item{N}{Integer. Number of elements to draw from.}

\item{n}{Integer. Number of elements in the sample.}
}
\value{
Vector of length \code{n} of integers from the range \code{0} to \code{N-1}.
}
\description{
Samples \code{n} integers from the range \code{0} to \code{N-1}.
A call \code{srswr_rcpp_zero(N, n)} is statistically equivalent to R's
\code{sample(N, n, replace = TRUE) - 1} however the results of every individual
call are different due to differences in the implementation.
\code{srswr_rcpp_zero} is up to 5 times faster.
}
\examples{
srswr_rcpp_zero(10, 20)

}
