% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UPpoisson_rcpp}
\alias{UPpoisson_rcpp}
\title{Poisson sampling}
\usage{
UPpoisson_rcpp(pik)
}
\arguments{
\item{pik}{Numeric vector. First-order inclusion probabilities.}
}
\value{
Integer vector of size \code{length(pik)} (population size) with
elements 1 (if the unit is selected in the sample) or 0 (if it is not).
}
\description{
Draws a Poisson sample (without replacement, random sample size) using a
given vector of first-order inclusion probabilities. Rewriten in Rcpp based
on the \verb{\link[sampling]\{UPpoisson\}} function.
}
\examples{
UPpoisson_rcpp((1:10) / 55 * 5)
UPpoisson_rcpp(inclusion_probabilities_rcpp(runif(10), 5))

}
